fix : 공통된 max 검증 로직 공통 핸들러로 분리
